add EsxiFile implementing AsyncRead
The file handle will be used to as "inodes" when the kernel does a
"lookup". The AsyncRead impl is so that we can use a BufReader with it
without accidentally downloading a GB-sized vmdk when scanning them
for whether they're a descriptive text or a blob.

Although we might not even need to scan them if qemu/pve don't need to
actually list the directory contents but rather just go from whatever
the vmdks contain anyway?

Signed-off-by: Wolfgang Bumiller <w.bumiller@proxmox.com>