deal with file size changes during reading
We cache file sizes, but if the vmx config files change they can of
course very well also change their size, in which case we might get an
http 416 Range Not Satisfiable response. Let's consider the File
handle to be at EOF at that point.

additionally, partial reads return a `content-range` header which
includes the file's size, so use this to keep track of it

Signed-off-by: Wolfgang Bumiller <w.bumiller@proxmox.com>