reformat let-else for better readability
`cargo fmt` still mostly ignores let-else, so format it manually in
how most examples use it, with the just the inside of the else branch
block on a new, indented line. As otherwise, with no indentation
change between the assignment and the else in a new line it reads like
being two separate statements.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>