version file: skip repo revision for now
The current git revision is not exposed by cargo, so we'd need a
build.rs script to get this, like we do in PBS (well in a sub-crate to
avoid frequent costly rebuilds). For now that's not worth it, having
the version exposed is enough to see if the mount is old.

And do not concat the three version atoms manually, we can use the
CARGO_PKG_VERSION environment variable.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>