esxi: add help and version flags as long and short option
even though this is in libexec and not meant for direct use it can be
still useful to get the version of the installed file to compare with
the one of a mounted FS, or to get the help output if one is testing
something manually.

So add the classic -v or --version and the -h or --help flags,
respectively.

Any of those will cause an early (successful) exit.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>