listvms: exclude VMs without datastore_name from manifest.json.
On newer ESXI-8.x.y versions, the diskless vCLS machines in an
ESXI-cluster are not stored on a datastore anymore.

Instead, they are placed under `/var/run/crx` on the ESXI-hosts'
filesystem. This can lead to issues with the ESXI-storage not being
activated on the PVE-side [0].

This commit changes the default behaviour of `manifest.json` creation by
excluding any machine that has an empty datastore-string in its config
and by that prevents querying a file that we cannot reach anyway.

[0] https://forum.proxmox.com/threads/new-import-wizard-available-for-migrating-vmware-esxi-based-virtual-machines.144023/post-759288

Signed-off-by: Daniel Herzig <d.herzig@proxmox.com>